requirements.txt
Flask
google-genai
google-generativeai